Install packages for linux
The script downloads the packages that are necessary for the sherlock.py. It is an automated procedure instead of the instructions in the README.md.
It contains some packages that are not in README but I couldn't run "pip3 install -r requirements.txt" without them.
It is for linux, I don't know if it works on something else.

I believe it is useful!