Make BANNER an r'string' and move it up front
Sherlock has a cool logo so this PR moves it to the top of sherlock.py to show it off when potential contributors read our source.  It also makes a BANNER an __r'string'__ so we do not need backslashes to escape characters and the ascii art is more what-you-see-is-what-you-get.  This also suppresses [a bunch of flake8 complaints](https://travis-ci.com/TheYahya/sherlock/jobs/172028552#L490-L503) about _invalid escape sequences_.  Finally this PR runs isort on the imports for readability.  Standard Lib imports come first, etc.