Change method used to get site ranking. Not only has alexa.com changed the format of their web site, they also seem to have gotten even more picky about people scraping it.  So, use their API to query the site data.
Not sure why the API was not used earlier.  It is much more efficient as much less data is passed.  And, it is less likely to have its format changed randomly.  It did flake out on some sites one time I used it.  But, then it worked reliably next.  I am not going to do many requests, as it probably has some query rate limit.