Fix Error When csv Is Written When Username Is Illegal For A Site
This problem happens during the write to the CSV.  There are various results that are gathered during the query for the existence of a given site, and they are always expected to exist in the results dictionary.  However, if a given user name has an illegal format for a given site, then the query is not even attempted.  Which leaves some keys in the dictionary undefined.

The solution is to define these keys even in the case where the username has an illegal format.