Limit Number Of Parallel Requests To 20 (Instead Of Number Of Sites)
Previous code was allocating room for as many workers as there was sites.  The problem is that as the number of sites has grown, there has not been enough memory to allocate all of those requests.  In reality, having all of these requests in parallel does not really speed the processing: on my computer, the time to do a query for all of the sites was 1 minute 10 seconds before the change, and was 1 minute 9 seconds after the change.

Limiting the number of workers to 10 did increase the query time to 1 minute 17s.  I am not sure if that is just inconsistencies in network traffic, but I will leave the limit at 20 for now.

Note that with the limit of 20, my query detected more sites than it did previously.  It appears that some of the requests were failing on my computer because of memory reasons (as opposed to actual detection on the site).