Restructure all print output to use QueryNotifyPrint() object.
Added start and finish methods to base QueryNotify() object in order to get the same type of output.

The only thing not supported in this refactor is the exception text for an error status when we are in verbose mode.  This is an area of future work: I think exception information like this would more properly be handled by the logging module.