Add override to HTTP Status Detection so HEAD request is not used.  Configure Instagram to use this option.
In most cases when we are detecting by status code, it is not necessary to get the entire body:  we can detect fine with just the HEAD response.  However, Richard Getz discovered that some sites (e.g. Instagram) will not respond properly if Sherlock only requests the HEAD.

Add a "request_head_only" attribute to the data.json so HTTP Status Detection can be configured either way.  It is simpler to support this change in this fashion, as it does not require changes to the tests.

With Richard Getz <richardgetziii@gmail.com>