Sherlock will from now on not use the local data.json
It will now use the data that is in the GitHub repo instead. The reason why we are using
this instead of the local one is so that the user has the most up to date data. This prevents
users from creating issue about false positives which has already been fixed or having outdated data