Print found only by default
As mentioned in #718, it would be more useful for the user of Sherlock
to only get the results of the sites that return a positive result. With
these new changes, if you want to all results to be printed out, then
you can do that by using the --verbose flag.