Update to data.jason
Added dangerousthings.com as a part of the list so that you are able to see if a "user" is interested in Biohacking. Just in case the forum is a tight community and very few people have interest in the site's offerings so I think it's a good thing to know about your target in an OSINT investigation of some sort.