Sherlock can now lookup usernames that begin with Hyphens
I've added more detailed instructions on how to lookup usernames starting with hyphens, as well as added some code to remove the trailing whitespace character in front of the hypen.

The only solution to this issue, without major changes to the current code, is to tell the user to incase the target username within quotations as well as prefix a space to the target username. For example ' -username'.

The reason why this has to be done is because it's a bug within argparse itself as can be seen here: https://bugs.python.org/issue9334

The conclusion of the thread was that it's not worth fixing and therefore it still remains.