Update README.md
In response to the issue, the proposed solution involves adding a link to the "sherlock-rest" repository in the README of the "sherlock-project/sherlock" repository. This link will provide users with information about an alternative REST+JSON endpoint for the "sherlock" project, enhancing the project's accessibility and offering users more options.