added new logo
The logo was created by the Kali Linux team and gave us permission to
use it.

In addition to the new logo, the ./images/ directory has been
moved to into ./docs/ in order to make the root directory of the project
cleaner. The old preview.png was never used as we had been using a
preview image that was hosted on user-images.githubuesrcontent.com. This
too has been changed so that we use the local copy instead